coolsite/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.10
django-celery-beat==2.5.0
django-debug-toolbar==4.3.0
django-fontawesome==1.0
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
flake8==7.0.0
fontawesome-free==5.15.4
fontawesomefree==6.5.1
isort==5.13.2
kombu==5.3.5
mccabe==0.7.0
mysqlclient==2.2.1
pillow==10.2.0
prompt-toolkit==3.0.43
pycodestyle==2.11.1
pyflakes==3.2.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
redis-server==6.0.9
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urlopen==1.0.0
users==1.0.dev0
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
